#5- Gracefully organize the folder structure for Storybooks
